Convert Form Button to Typescript
