--- src/features/app/components/Form/Button.js
@@ -1,32 +1,0 @@
-import { forwardRef } from 'react';
-import PropTypes from 'prop-types';
-
-import Loading from '../Loading';
-
-import { Button, loadingStyles } from './Form.styles';
-
-const FormButton = forwardRef(
-  ({ isDisabled, isLoading, text, ...rest }, ref) => (
-    <Button
-      ref={ref}
-      type="submit"
-      disabled={isDisabled || isLoading}
-      {...rest}
-    >
-      {isLoading ? <Loading styles={loadingStyles} /> : text}
-    </Button>
-  )
-);
-
-FormButton.defaultProps = {
-  isDisabled: false,
-  isLoading: false,
-};
-
-FormButton.propTypes = {
-  isDisabled: PropTypes.bool,
-  isLoading: PropTypes.bool,
-  text: PropTypes.string.isRequired,
-};
-
-export default FormButton;
